Supports headless configuration for auth button
