+++ package.json
@@ -57,8 +57,8 @@
     "@fontsource/quattrocento-sans": "^4.5.9",
     "@fractalwagmi/fractal-auth-api": "^0.0.22",
     "@fractalwagmi/fractal-auth-private-web-sdk-api": "^0.0.1",
-    "@fractalwagmi/fractal-sdk-websdk-api": "^0.0.1",
     "@fractalwagmi/fractal-sdk-api": "^0.0.5",
+    "@fractalwagmi/fractal-sdk-websdk-api": "^0.0.1",
     "jest": "^28.1.3",
     "jwt-decode": "^3.1.2",
     "swr": "^1.3.0"
