+++ src/hooks/public/index.ts
@@ -1,7 +1,8 @@
+export * from 'hooks/public/use-auth-button-props';
 export * from 'hooks/public/use-coins';
 export * from 'hooks/public/use-items';
-export * from 'hooks/public/use-user-wallet';
-export * from 'hooks/public/use-user';
+export * from 'hooks/public/use-sign-out';
 export * from 'hooks/public/use-sign-out';
 export * from 'hooks/public/use-sign-transaction';
-export * from 'hooks/public/use-sign-out';
+export * from 'hooks/public/use-user-wallet';
+export * from 'hooks/public/use-user';
